pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "q-uestionable-ai"
version = "0.1.0"
description = "Offensive security platform for agentic AI infrastructure"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Richard Spicer", email = "richard@richardspicer.io" },
]
keywords = ["security", "ai", "mcp", "red-team", "offensive-security"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Security",
]
dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "pyyaml>=6.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2>=3.1",
    "websockets>=15.0",
    "mcp>=1.26,<2",
    "anthropic>=0.84",
]

[project.scripts]
qai = "q_ai.cli:app"

[project.urls]
Homepage = "https://q-uestionable.ai"
Repository = "https://github.com/q-uestionable-AI/qai"
Issues = "https://github.com/q-uestionable-AI/qai/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/q_ai"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-timeout>=2.2",
    "pre-commit>=3.5",
    "mypy>=1.8",
    "ruff>=0.4",
    "types-pyyaml>=6.0",
    "httpx>=0.27",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "S", "A", "C4", "DTZ", "T20", "RUF"]
ignore = [
    "S101",  # assert usage (fine in tests)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S"]
"src/q_ai/**/cli*.py" = ["B008"]  # typer.Option() in defaults is the standard pattern
"src/q_ai/core/db.py" = ["A002"]  # type/hash params match DB column names
"src/q_ai/server/**" = ["B008"]  # FastAPI Depends() in defaults

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_untyped_defs = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
timeout = 30
markers = [
    "integration: marks tests requiring external services (deselect with '-m \"not integration\"')",
]
